# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/v1/generation/text-to-image": {
            "post": {
                "summary": "Text2Img Generation",
                "operationId": "text2img_generation_v1_generation_text_to_image_post",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes",
                            "title": "Accept"
                        },
                        "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes"
                    },
                    {
                        "name": "accept",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Accept"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Text2ImgRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PNG bytes if request's 'Accept' header is 'image/png', otherwise JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GeneratedImageBase64"
                                    },
                                    "title": "Response Text2Img Generation V1 Generation Text To Image Post"
                                },
                                "example": [
                                    {
                                        "base64": "...very long string...",
                                        "seed": 1050625087,
                                        "finish_reason": "SUCCESS"
                                    }
                                ]
                            },
                            "image/png": {
                                "example": "PNG bytes, what did you expect?"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/generation/image-upscale-vary": {
            "post": {
                "summary": "Img Upscale Or Vary",
                "operationId": "img_upscale_or_vary_v1_generation_image_upscale_vary_post",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes",
                            "title": "Accept"
                        },
                        "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes"
                    },
                    {
                        "name": "accept",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Accept"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_img_upscale_or_vary_v1_generation_image_upscale_vary_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PNG bytes if request's 'Accept' header is 'image/png', otherwise JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GeneratedImageBase64"
                                    },
                                    "title": "Response Img Upscale Or Vary V1 Generation Image Upscale Vary Post"
                                },
                                "example": [
                                    {
                                        "base64": "...very long string...",
                                        "seed": 1050625087,
                                        "finish_reason": "SUCCESS"
                                    }
                                ]
                            },
                            "image/png": {
                                "example": "PNG bytes, what did you expect?"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/generation/image-inpait-outpaint": {
            "post": {
                "summary": "Img Inpaint Or Outpaint",
                "operationId": "img_inpaint_or_outpaint_v1_generation_image_inpait_outpaint_post",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes",
                            "title": "Accept"
                        },
                        "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes"
                    },
                    {
                        "name": "accept",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Accept"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_img_inpaint_or_outpaint_v1_generation_image_inpait_outpaint_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PNG bytes if request's 'Accept' header is 'image/png', otherwise JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GeneratedImageBase64"
                                    },
                                    "title": "Response Img Inpaint Or Outpaint V1 Generation Image Inpait Outpaint Post"
                                },
                                "example": [
                                    {
                                        "base64": "...very long string...",
                                        "seed": 1050625087,
                                        "finish_reason": "SUCCESS"
                                    }
                                ]
                            },
                            "image/png": {
                                "example": "PNG bytes, what did you expect?"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AspectRatio": {
                "type": "string",
                "enum": [
                    "704×1408",
                    "704×1344",
                    "768×1344",
                    "768×1280",
                    "832×1216",
                    "832×1152",
                    "896×1152",
                    "896×1088",
                    "960×1088",
                    "960×1024",
                    "1024×1024",
                    "1024×960",
                    "1088×960",
                    "1088×896",
                    "1152×896",
                    "1152×832",
                    "1216×832",
                    "1280×768",
                    "1344×768",
                    "1344×704",
                    "1408×704",
                    "1472×704",
                    "1536×640",
                    "1600×640",
                    "1664×576",
                    "1728×576"
                ],
                "title": "AspectRatio"
            },
            "Body_img_inpaint_or_outpaint_v1_generation_image_inpait_outpaint_post": {
                "properties": {
                    "input_image": {
                        "type": "string",
                        "format": "binary",
                        "title": "Input Image"
                    },
                    "input_mask": {
                        "type": "string",
                        "format": "binary",
                        "title": "Input Mask"
                    },
                    "outpaint_selections": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Outpaint Selections",
                        "default": []
                    },
                    "prompt": {
                        "type": "string",
                        "title": "Prompt",
                        "default": ""
                    },
                    "negative_promit": {
                        "type": "string",
                        "title": "Negative Promit",
                        "default": ""
                    },
                    "style_selections": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Style Selections",
                        "default": [
                            "Fooocus V2",
                            "Default (Slightly Cinematic)"
                        ]
                    },
                    "performance_selection": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PerfomanceSelection"
                            }
                        ],
                        "default": "Speed"
                    },
                    "aspect_ratios_selection": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AspectRatio"
                            }
                        ],
                        "default": "1152×896"
                    },
                    "image_number": {
                        "type": "integer",
                        "maximum": 32,
                        "minimum": 1,
                        "title": "Image Number",
                        "description": "Image number",
                        "default": 1
                    },
                    "image_seed": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Seed"
                    },
                    "sharpness": {
                        "type": "number",
                        "maximum": 30,
                        "minimum": 0,
                        "title": "Sharpness",
                        "default": 2
                    },
                    "guidance_scale": {
                        "type": "number",
                        "maximum": 30,
                        "minimum": 1,
                        "title": "Guidance Scale",
                        "default": 7
                    },
                    "base_model_name": {
                        "type": "string",
                        "title": "Base Model Name",
                        "default": "sd_xl_base_1.0_0.9vae.safetensors"
                    },
                    "refiner_model_name": {
                        "type": "string",
                        "title": "Refiner Model Name",
                        "default": "sd_xl_refiner_1.0_0.9vae.safetensors"
                    },
                    "l1": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L1",
                        "default": "sd_xl_offset_example-lora_1.0.safetensors"
                    },
                    "w1": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W1",
                        "default": 0.5
                    },
                    "l2": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L2"
                    },
                    "w2": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W2",
                        "default": 0.5
                    },
                    "l3": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L3"
                    },
                    "w3": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W3",
                        "default": 0.5
                    },
                    "l4": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L4"
                    },
                    "w4": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W4",
                        "default": 0.5
                    },
                    "l5": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L5"
                    },
                    "w5": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W5",
                        "default": 0.5
                    }
                },
                "type": "object",
                "required": [
                    "input_image"
                ],
                "title": "Body_img_inpaint_or_outpaint_v1_generation_image_inpait_outpaint_post"
            },
            "Body_img_upscale_or_vary_v1_generation_image_upscale_vary_post": {
                "properties": {
                    "input_image": {
                        "type": "string",
                        "format": "binary",
                        "title": "Input Image"
                    },
                    "uov_method": {
                        "$ref": "#/components/schemas/UpscaleOrVaryMethod"
                    },
                    "prompt": {
                        "type": "string",
                        "title": "Prompt",
                        "default": ""
                    },
                    "negative_promit": {
                        "type": "string",
                        "title": "Negative Promit",
                        "default": ""
                    },
                    "style_selections": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Style Selections",
                        "default": [
                            "Fooocus V2",
                            "Default (Slightly Cinematic)"
                        ]
                    },
                    "performance_selection": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PerfomanceSelection"
                            }
                        ],
                        "default": "Speed"
                    },
                    "aspect_ratios_selection": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AspectRatio"
                            }
                        ],
                        "default": "1152×896"
                    },
                    "image_number": {
                        "type": "integer",
                        "maximum": 32,
                        "minimum": 1,
                        "title": "Image Number",
                        "description": "Image number",
                        "default": 1
                    },
                    "image_seed": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Seed"
                    },
                    "sharpness": {
                        "type": "number",
                        "maximum": 30,
                        "minimum": 0,
                        "title": "Sharpness",
                        "default": 2
                    },
                    "guidance_scale": {
                        "type": "number",
                        "maximum": 30,
                        "minimum": 1,
                        "title": "Guidance Scale",
                        "default": 7
                    },
                    "base_model_name": {
                        "type": "string",
                        "title": "Base Model Name",
                        "default": "sd_xl_base_1.0_0.9vae.safetensors"
                    },
                    "refiner_model_name": {
                        "type": "string",
                        "title": "Refiner Model Name",
                        "default": "sd_xl_refiner_1.0_0.9vae.safetensors"
                    },
                    "l1": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L1",
                        "default": "sd_xl_offset_example-lora_1.0.safetensors"
                    },
                    "w1": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W1",
                        "default": 0.5
                    },
                    "l2": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L2"
                    },
                    "w2": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W2",
                        "default": 0.5
                    },
                    "l3": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L3"
                    },
                    "w3": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W3",
                        "default": 0.5
                    },
                    "l4": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L4"
                    },
                    "w4": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W4",
                        "default": 0.5
                    },
                    "l5": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "L5"
                    },
                    "w5": {
                        "type": "number",
                        "maximum": 2,
                        "minimum": -2,
                        "title": "W5",
                        "default": 0.5
                    }
                },
                "type": "object",
                "required": [
                    "input_image",
                    "uov_method"
                ],
                "title": "Body_img_upscale_or_vary_v1_generation_image_upscale_vary_post"
            },
...
            "GeneratedImageBase64": {
                "properties": {
                    "base64": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Base64",
                        "description": "Image encoded in base64, or null if finishReasen is not 'SUCCESS'"
                    },
                    "seed": {
                        "type": "integer",
                        "title": "Seed",
                        "description": "The seed associated with this image"
                    },
                    "finish_reason": {
                        "$ref": "#/components/schemas/GenerationFinishReason"
                    }
                },
                "type": "object",
                "required": [
                    "base64",
                    "seed",
                    "finish_reason"
                ],
                "title": "GeneratedImageBase64"
            },
            "GenerationFinishReason": {
                "type": "string",
                "enum": [
                    "SUCCESS",
                    "QUEUE_IS_FULL",
                    "USER_CANCEL",
                    "ERROR"
                ],
                "title": "GenerationFinishReason"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PerfomanceSelection": {
                "type": "string",
                "enum": [
                    "Speed",
                    "Quality"
                ],
                "title": "PerfomanceSelection"
            },
            "Text2ImgRequest": {
                "properties": {
                    "prompt": {
                        "type": "string",
                        "title": "Prompt",
                        "default": ""
                    },
                    "negative_promit": {
                        "type": "string",
                        "title": "Negative Promit",
                        "default": ""
                    },
                    "style_selections": {
                        "items": {
                            "$ref": "#/components/schemas/FooocusStyle"
                        },
                        "type": "array",
                        "title": "Style Selections",
                        "default": [
                            "Fooocus V2",
                            "Default (Slightly Cinematic)"
                        ]
                    },
                    "performance_selection": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PerfomanceSelection"
                            }
                        ],
                        "default": "Speed"
                    },
                    "aspect_ratios_selection": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AspectRatio"
                            }
                        ],
                        "default": "1152×896"
                    },
                    "image_number": {
                        "type": "integer",
                        "title": "Image Number",
                        "description": "Image number",
                        "default": 1,
                        "max": 32,
                        "min": 1
                    },
                    "image_seed": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Seed"
                    },
                    "sharpness": {
                        "type": "number",
                        "title": "Sharpness",
                        "default": 2,
                        "max": 30,
                        "min": 0
                    },
                    "guidance_scale": {
                        "type": "number",
                        "title": "Guidance Scale",
                        "default": 7,
                        "max": 30,
                        "min": 1
                    },
                    "base_model_name": {
                        "type": "string",
                        "title": "Base Model Name",
                        "default": "sd_xl_base_1.0_0.9vae.safetensors"
                    },
                    "refiner_model_name": {
                        "type": "string",
                        "title": "Refiner Model Name",
                        "default": "sd_xl_refiner_1.0_0.9vae.safetensors"
                    },
                    "loras": {
                        "items": {
                            "$ref": "#/components/schemas/Lora"
                        },
                        "type": "array",
                        "maxItems": 5,
                        "title": "Loras",
                        "default": [
                            {
                                "model_name": "sd_xl_offset_example-lora_1.0.safetensors",
                                "weight": 0.5
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "Text2ImgRequest"
            },
...
        }
    }
}